Semi-working code

+++ src/tasks/rm3100/rm3100.h
@@ -95,9 +95,9 @@ void rm3100_main(void *pvParameters);
 int init_rm3100(void);
 
 typedef struct {
-    long x;
-    long y;
-    long z;
+    uint32_t x;
+    uint32_t y;
+    uint32_t z;
 } RM3100_return_t;
 
 typedef enum {
